LayoutTests/platform/win-future/accessibility/img-fallsback-to-title-expected.txt

    
test
test
This tests that images will fallback to using the title attribute if no other descriptive text is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Image1 description: Description: test1
Image1 help: 

Image2 description: Description: alt
Image2 help: 

Image3 description: Description: test3
Image3 help: 

Image4 description: Description: alt
Image4 help: 

PASS imagesGroup.childAtIndex(0).childrenCount is 2
PASS successfullyParsed is true

TEST COMPLETE
